Moved some constants from riskmodels->scientific
